% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2matR.R
\name{df2matR}
\alias{df2matR}
\title{Transform Dataframe to Matrix R}
\usage{
df2matR(var.df, r)
}
\arguments{
\item{var.df}{dataframe of sampling variances of direct estimators.}

\item{r}{number of variabel}
}
\value{
Block diagonal matrix R
}
\description{
This function transforms dataframe contains sampling variance to block diagonal matrix R
}
\examples{
NULL
}
